# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: contact@josehernandez.dev
    name: Smidgen
    url: https://example.com/
  description: API for interacting with Smidgen.
  license:
    name: " GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007"
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: Smidgen
  version: 0.0.1
servers:
- url: /api/v1
paths:
...
  /business_unit/:
    get:
      description: Get all business units stored in the database.
      operationId: get_unit_business_unit__get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/business_unit'
                title: Response Get Unit Business Unit  Get
                type: array
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
      summary: Get Business Units
      tags:
      - business unit
    post:
      description: Create a new business unit in the database.
      operationId: add_unit_business_unit__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/business_unit'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Business Unit
      tags:
      - business unit
  /business_unit/{unit_id}:
    delete:
      description: Delete the specified business unit from the database.
      operationId: delete_unit_business_unit__unit_id__delete
      parameters:
      - explode: false
        in: path
        name: unit_id
        required: true
        schema:
          title: Unit Id
          type: integer
        style: simple
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Business Unit
      tags:
      - business unit
    get:
      description: Get the specified business unit stored in the database.
      operationId: get_units_business_unit__unit_id__get
      parameters:
      - explode: false
        in: path
        name: unit_id
        required: true
        schema:
          title: Unit Id
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/business_unit'
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Business Unit
      tags:
      - business unit
    put:
      description: Update the specified business unit in the database.
      operationId: update_unit_business_unit__unit_id__put
      parameters:
      - explode: false
        in: path
        name: unit_id
        required: true
        schema:
          title: Unit Id
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/business_unit'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Business Unit
      tags:
      - business unit
...
components:
  schemas:
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      example:
        msg: msg
        loc:
        - Location_inner
        - Location_inner
        type: type
      properties:
        loc:
          items:
            $ref: '#/components/schemas/Location_inner'
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    business_unit:
      example:
        country: country
        city: city
        address_line_two: address_line_two
        name: name
        state: state
        address_line_one: address_line_one
        unit_id: 0
        point_of_contact: point_of_contact
      properties:
        unit_id:
          title: Unit Id
          type: integer
        name:
          title: Name
          type: string
        point_of_contact:
          title: Point Of Contact
          type: string
        address_line_one:
          title: Address Line One
          type: string
        address_line_two:
          title: Address Line Two
          type: string
        state:
          title: State
          type: string
        city:
          title: City
          type: string
        country:
          title: Country
          type: string
      required:
      - address_line_one
      - address_line_two
      - city
      - country
      - name
      - point_of_contact
      - state
      - unit_id
      title: business_unit
      type: object
...
    Location_inner:
      anyOf:
      - type: string
      - type: integer